# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Sentinel AWS - Cloud Security Auditing Tool API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /scans:
    get:
      operationId: listScans
      tags: [scans]
      summary: List scan history
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: accountId
          schema:
            type: number
        - in: query
          name: page
          schema:
            type: number
        - in: query
          name: limit
          schema:
            type: number
      responses:
        "200":
          description: Scan list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScanListResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    post:
      operationId: startScan
      tags: [scans]
      summary: Start a new security scan
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StartScanRequest"
      responses:
        "202":
          description: Scan started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Scan"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /scans/{id}:
    get:
      operationId: getScan
      tags: [scans]
      summary: Get scan details
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Scan details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScanDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /scans/{id}/compare/{compareWithId}:
    get:
      operationId: compareScan
      tags: [scans]
      summary: Compare scan with a previous scan
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: number
        - in: path
          name: compareWithId
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Comparison result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScanComparison"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # Findings
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
        - message

    # Auth
...
    StartScanRequest:
      type: object
      properties:
        accountId:
          type: number
        modules:
          type: array
          items:
            type: string
            enum: [iam, s3, ec2, security_groups, cloudtrail]
      required:
        - accountId
    Scan:
      type: object
      properties:
        id:
          type: number
        accountId:
          type: number
        accountName:
          type: string
        status:
          type: string
          enum: [pending, running, completed, failed]
        progress:
          type: number
        currentModule:
          type: ["string", "null"]
        overallScore:
          type: ["number", "null"]
        iamScore:
          type: ["number", "null"]
        s3Score:
          type: ["number", "null"]
        ec2Score:
          type: ["number", "null"]
        networkingScore:
          type: ["number", "null"]
        cloudtrailScore:
          type: ["number", "null"]
        criticalCount:
          type: number
        highCount:
          type: number
        mediumCount:
          type: number
        lowCount:
          type: number
        startedAt:
          type: string
        completedAt:
          type: ["string", "null"]
        errorMessage:
          type: ["string", "null"]
      required:
        - id
        - accountId
        - accountName
        - status
        - progress
        - criticalCount
        - highCount
        - mediumCount
        - lowCount
        - startedAt
    ScanDetail:
      allOf:
        - $ref: "#/components/schemas/Scan"
        - type: object
          properties:
            findings:
              type: array
              items:
                $ref: "#/components/schemas/Finding"
          required:
            - findings
    ScanListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Scan"
        total:
          type: number
        page:
          type: number
        limit:
          type: number
        totalPages:
          type: number
      required:
        - data
        - total
        - page
        - limit
        - totalPages
    ScanComparison:
      type: object
      properties:
        current:
          $ref: "#/components/schemas/Scan"
        previous:
          $ref: "#/components/schemas/Scan"
        scoreDelta:
          type: number
        newFindings:
          type: array
          items:
            $ref: "#/components/schemas/Finding"
        resolvedFindings:
          type: array
          items:
            $ref: "#/components/schemas/Finding"
        persistingFindings:
          type: array
          items:
            $ref: "#/components/schemas/Finding"
      required:
        - current
        - previous
        - scoreDelta
        - newFindings
        - resolvedFindings
        - persistingFindings

    # Findings
    Finding:
      type: object
      properties:
        id:
          type: number
        scanId:
          type: number
        service:
          type: string
          enum: [iam, s3, ec2, security_groups, cloudtrail]
        checkId:
          type: string
        title:
          type: string
        description:
          type: string
        severity:
          type: string
          enum: [critical, high, medium, low, informational]
        affectedResource:
          type: string
        recommendation:
          type: string
        metadataJson:
          type: ["string", "null"]
      required:
        - id
        - scanId
        - service
        - checkId
        - title
        - description
        - severity
        - affectedResource
        - recommendation
...