# lib/api-zod/src/generated/types/startScanRequest.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Sentinel AWS - Cloud Security Auditing Tool API
 * OpenAPI spec version: 0.1.0
 */
import type { StartScanRequestModulesItem } from './startScanRequestModulesItem';

export interface StartScanRequest {
  accountId: number;
  modules?: StartScanRequestModulesItem[];
}